2.3 commit
I tried to make a flip animation on the card page; it's still a work in
progress, but it's late so I'll finish later. This update simply has the
card activity added so you can work on other things.

The main thing after the flip is the 'filter' menu - perhaps we can use
a material design-y dialog box over a blurred screen?